spectrum_server/pyproject.toml
[project]
name = "spectrum_server"
version = "1.1.0"
description = "MCP-enabled spectrum analysis server for Ettus B210 SDR with GPU-accelerated processing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.116.1",
    "fastmcp==2.10.6",
    "pydantic==2.11.7",
    "uvicorn>=0.35.0",
    "pyzmq>=26.0.0",
    "pymongo",
    "psutil",
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    "nvidia-ml-py>=12.555.43",
]

[project.optional-dependencies]
# GPU acceleration with CuPy (requires CUDA)
gpu = [
    "cupy-cuda12x>=13.0.0",
]
# UHD support for Ettus devices
uhd = [
    # uhd Python bindings installed via system package
]
# SoapySDR support
soapy = [
    # SoapySDR Python bindings installed via system package
]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
    "mypy",
    "types-psutil",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
